templates/template-data-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-pipeline"
version = "0.1.0"
description = "Data pipeline template"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas==2.1.4",
    "pyyaml==6.0.1",
    "pydantic==2.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "ruff==0.1.13",
]

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
